requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncpg==0.30.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dictdiffer==0.9.0
ecdsa==0.19.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
iso8601==2.1.0
mypy-protobuf==3.6.0
protobuf==5.29.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pypika-tortoise==0.5.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tortoise-orm==0.24.2
types-protobuf==5.29.1.20250403
types-pyasn1==0.6.0.20250208
types-python-jose==3.4.0.20250224
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
